fix: relocate stopping criteria to a module script.
next step is to expant the stopping criteria such that it covers many other.
